requirements.txt
asgiref==3.7.2
Django==3.2.20
gunicorn==21.2.0
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
djangorestframework==3.14.0